Move `Body` to the body mobule.
Add more comments.
